conat tests -- refactoring and improving
